requirements.txt
#fastapi
fastapi[standard]
uvicorn[standard]
sqlalchemy
pydantic-settings
python-dotenv
pymysql
cryptography
geoalchemy2
shapely
python-dateutil
pandas
sqlmodel